requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
mysqlclient==2.1.1
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
